bug encountered, temporal patch.
Encountered  a bug in which the randomPathGenerator loops infinitely, added a brute force exit with a counter. Source of the bug unknown, most probably originated with the use of the random number generator, given that the function has very strict validations. Using on graphs smaller than 15 elements is not adviced.